Streaming support
This modifies the code to allow reading from a pipe or socket, avoiding
the requirement to buffer the entire package in a file.  O_TMPFILE is
used on Linux to prevent untrusted data from being exposed.  On other
platforms, a temporary file with a name ending in .UNTRUSTED is used.