Suppress warnings on various versions of Rust
rpm-oxide needs to build on Rust as old as 1.21.0, which is missing many
modern features and had not yet stabilized others.  Many of the
workarounds result in warnings on modern compilers.

Fortunately, this can be solved by using a build script that
detects the version of ‘rustc’ and sets configuration options
accordingly.  These configuration options disable warnings on new
compilers, and enable not-yet-stabilized features on old compilers.  The
bulid script also sets the ‘RUSTC_BOOTSTRAP’ environment variable to 1
on old compilers, enabling the needed nightly features.  This would
result in an error with modern Cargo, but new Cargo + old rustc is not a
configuration we need to support.