Always produce new-format packets
This is RECOMMENDED by RFC4880 and should be supported by all versions
of RPM, as they use the same code for parsing new-format packets that
they use for parsing v4 signature subpackets.

This also fixes a nasty bug in the new-format packet generation code: if
the packet was longer than 8383 bytes, the second byte of the new packet
(0xFF) would be omitted, causing a corrupt packet to be produced.
This would have been misinterpreted by RPM.  Add a test case to ensure
that this does not happen.  Since signatures normally are nowhere near
this long, this bug could have escaped into the wild.