Serialize all calls to RPM keyring and transaction functions
It turns out that RPM’s keyring and transaction functions are not
thread-safe.  This showed up as crashes in RPM’s atexit() handler when
running the test suite, which were very difficult to debug.  Fix the
problem by wrapping all of these calls in a global lock.  Performance
should not be noticibly impacted, as the hot-path is the symmetric
cryptography and that can still happen in parallel.  rpmcanon is
single-threaded and so is not affected.

Tested by running the test suite several times on a multi-core system.
Without this patch a substantial fraction of runs will crash with
SIGSEGV or SIGABRT.  I cannot reproduce the crash with this patch
applied.