Remove a huge amount of unnecessary locking
Panu Matilainen (the RPM project lead) suggested loading the RPM keyring
eagerly instead of lazily.  This turns out to work really well: it
prevents a panic if the keyring fails to load and also improves
thread-safety.

He also mentioned that RPM keyrings are thread-safe and librpm handles
all needed synchronization internally.  A quick look at the 4.14.3
source code indicates that this is in fact the case, and so all of the
synchronization around keyring operations can go away.  As a precaution,
all operations on a transaction set are still serialized against both
each other and process exit.