Check that i18nstrings are valid and consistent
RPM assumes that the i18ntable is at least as long as all i18nstring
entries, and many versions will read out of bounds if the i18ntable is
too short.  Reject packages that have any of the following:

- An i18ntable entry or any i18nstring entries in the signature header.
- An i18ntable entry with a type other than RPM_STRING_ARRAY_TYPE,
  which would be unusable.
- An i18nstring entry with a greater length than the preceeding
  i18ntable.

However, an i18nstring entry with a smaller length than the preceeding
i18ntable is valid and must be accepted.  This just means that some
translations are missing, which is a valid and legitimate situation.
Some debuginfo packages have i18nstring entries without any i18ntable
entries being present, so allow this.

Tests for all of these cases are included.  The code is also reformatted
with `cargo fmt`.